requirements.txt
flask>=3.0.0
scenedetect[opencv]>=0.6.0
python-dotenv>=1.0.0
